requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
Werkzeug
